add endpoint for signup page and make separate file for signIn
