+++ index.js
@@ -1,14 +1,22 @@
 const express = require("express");
 const app = express();
 const cors = require("cors");
-app.use(cors());
+
 const jsonParser = express.json();
 app.use(jsonParser);
-const User = require("./routers/user");
+app.use(cors());
+const corsMiddleware =cors();
 
 const PORT = 5000;
+
+
 app.listen(PORT, () => {
   console.log("Listening on port");
 });
 
-app.use("/", User);
+//routers
+const User = require("./routers/user");
+app.use("/signup", User);
+
+const userLogin = require("./routers/auth")
+app.use("/login", userLogin)
